### 2025-06-25 #8489 @editorialbot
[PRE REVIEW]: matlab-igraph: bringing igraph to MATLAB
**Submitting author:** @DavidRConnell (<a href="http://orcid.org/0000-0002-4841-6756">David R. Connell</a>)
**Repository:** https://github.com/DavidRConnell/matlab-igraph
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.0
