### 2025-06-25 #8489 @DavidRConnell
Anything under `vendor` (i.e. `igraph`) is not original work as well as everything under `docs` other than `docs/content`. Everything else should be with the main contributions being under `toolbox`, `mex`, and `mxIgraph` and `tests` / `benchmarks` for tests.
