### 2025-06-25 #8489 @szhorvat
Keeping track of this, and happy to provide some input, but I believe I can't be a reviewer due to being an igraph maintainer.
