### 2025-07-03 #8489 @danielskatz
👋 @onurvarol & @blbentley & @sa501428 & @pausz & @LCB0B & @i64 - Can I check with you again to see if you would be able to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
