### 2025-07-08 #8489 @Jacob-Barhak
Sorry Daniel,

Your policy of not supporting licenses such as CC0 is problematic. I tried
to resolve it with NumFocus with no success. Since I cannot support your
policy and the name of the journal is actually misleading in my opinion
then I care not to contribute to its function.

In fact I use your journal as a negative example of how open source
communities misbehave with regards to openness. You can see references to
your journal and Numfocus in these talks:

   -

   Jacob Barhak, AI and Legal Aspects,
   https://www.clinicalunitmapping.com/show/AI_legal_Latest.pdf
   <https://www.clinicalunitmapping.com/show/AI_legal_Latest.pdf>
   -

      Austin Less Wrong, 2025 May 10th, Austin, TX
      -

        Austin Python Meetup, 2025 March 12th
      -

          Video: https://youtu.be/e19cYzXR0aA?si=8D7pXwUSG5Cx4Blv
         -

      Anaconda, 2025 Jan 9th, 17th
      -

      GE Healthcare Paper Club 2024 Sep 2nd Haifa
      -

      AI Makerspace Community Session 2024 Aug 24th, Sep 4th Virtual
      -

         Video 1:
         https://www.loom.com/share/a90ee4974a1643a9afebcd66122db510?sid=8de5eebd-6c56-415f-8048-a75117545f13
         -

         Video 2:
         https://www.loom.com/share/f6a0adcd0f88466392aeba897a189bc3?sid=ef4e5920-dd79-4219-819b-a427c09b1ef4

   -

   Jacob Barhak, "Open Source, Sustainability, and CC0" August Penguin
   3-Mar-2023, Tel Aviv, Israel. Presentation:
   https://jacob-barhak.github.io/PenguinIsrael2023_OpenSource_2023_03_02.pdf
   , Video: https://youtu.be/Bhli5L4qskU

I do provide reviews to many sources as you can see here:
https://groups.google.com/forum/#!forum/public-scientific-reviews

However, unless you change your policy with regards to the definition of
open source, I cannot honorably support your cause.

I apologize.

               Jacob

On Mon, Jul 7, 2025 at 7:44 AM Daniel S. Katz ***@***.***>
wrote:

> *danielskatz* left a comment (openjournals/joss-reviews#8489)
> <https://github.com/openjournals/joss-reviews/issues/8489#issuecomment-3044912488>
>
> 👋 @giovannivolpe <https://github.com/giovannivolpe> & @egolol
> <https://github.com/egolol> & @annacanal <https://github.com/annacanal> &
> @c-yuwei <https://github.com/c-yuwei> - based in part on your work on
> BRAPH, would one or more of you be willing to review this submission for
> JOSS? We carry out our checklist-driven reviews here in GitHub issues and
> follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
