### 2025-07-10 #8489 @danielskatz
@DavidRConnell - as I think I mentioned earlier, finding reviewers for Matlab submission is difficult. Do you have any other suggestions?

Also, I will mention that I don't think the conflict that @szhorvat mentioned is actually a conflict, as a developer of software used by your software is well positioned to improve your software to the acceptable point (if this is needed). They have said elsewhere that they could review this, starting in August, so that's a fallback position if we can't find anyone else before then.

And @gartavanis - sorry this is taking so long to get started.
