### 2025-07-10 #8489 @DavidRConnell
@danielskatz, I appreciate the effort. The use of MATLAB might not really fit the spirit of FOSS. I had seen the gramm MATLAB had been published here so I thought it would work but can look elsewhere if there aren't editors available.
