### 2025-07-11 #8489 @danielskatz
@DavidRConnell - the challenge isn't available editors, as that's me in this case, but willing reviewers, which is where we are failing.

The options I see are:
1. you provide more suggestions
2. we wait until August when @szhorvat would be willing to review
3. we give up

Which would you like to do?
