### 2025-07-11 #8489 @bencardoen
@danielskatz perhaps not my place to comment as I've declined to review, but perhaps there's an option 4? If the project/code can run under Octave, would that allow for both a larger audience, including those without access, and address the open/closed discussion point?
I don't know how much effort would be involved, but perhaps it is worth considering?
I haven't used Octave in a while, so I don't know what the current compatibilities would be, but I figured it can't hurt to mention it.
