### 2025-07-11 #8489 @DavidRConnell
@bencardoen for now I'm leaning towards give up. I didn't expect it would be this much a challenge.

@bencardoen That's possible. Last time I tried Octave it wasn't that straight forward to get my matlab code running on it (although that may have been my inexperience) but I can give it a chance once I finished with some other things. I would much prefer to not have my code dependent on a licensed language so that would be nice if it doesn't require too much change.
