### 2025-07-20 #8489 @platipodium
@danielskatz I would be able to review Matlab as I have a licensed version.  But @DavidRConnell an open-source platform like octave would be much preferable.  I strongly urge you to test disclose (in)-compatibilities in your documentation
