### 2025-07-20 #8622 @editorialbot
[REVIEW]: matlab-igraph: bringing igraph to MATLAB
**Submitting author:** @DavidRConnell (<a href="http://orcid.org/0000-0002-4841-6756">David R. Connell</a>)
**Repository:** https://github.com/DavidRConnell/matlab-igraph
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.0
