### 2025-07-21 #8489 @DavidRConnell
@platipodium thank you. And I agree, I would much prefer all my work to run without the need of licensed code in general. It's been a while since I have even looked at Octave and I don't know anything about its mex equivalent, but I will give it a try to see how feasible it looks. It will be at least a few weeks before I get a chance though.
