### 2025-07-21 #8489 @Jacob-Barhak
So Daniel,

This shows my point. JOSS has the wrong rule in place and it cannot change
it.

And @danielskatz, releasing the same code under multiple licenses is
allowed if one owns the copyright. So why restrict the developer?

The developer can potentially release the code under one license and at the
same time waive copyright under CC0 to give the code more freedom.

This solution will allow publication under your rules and more freedom
otherwise.

      Jacob
