### 2025-07-22 #8489 @DavidRConnell
@Jacob-Barhak I wrote my last comment incorrectly. When I said "without the need of licensed code", I was referring to the need for a license to access the code (i.e. buying a MATLAB license), which, based on this thread alone, clearly causes barriers that make my code less useful.

When choosing my license, I explicitly want my code to be "copyleft." I want my code to be open and I want it to remain open. I do not like the idea of someone being able to improve my code and then share it in a closed manner. There's a valid argument that this is less free than public domain but I think there is also a valid argument that it is more free. In the end, I will choose the license that is most likely to cause Steve Ballmer to cry about poor microsoft not being allowed to modify then sell my code.
