### 2025-07-22 DavidRConnell/matlab-igraph#4 @platipodium
Linking CXX shared module mexIgraphDispatcher.mexmaca64 fails under M1/arm64/clang
Related to JOSS review https://github.com/openjournals/joss-reviews/issues/8622#issuecomment-3094739199

I am following the installation instructions and need to expand the cmake options to the following:

```
cmake .. -DCMAKE_C_COMPILER=${CONDA_PREFIX}/bin/clang -DCMAKE_CXX_COMPILER=${CONDA_PREFIX}/bin/clang -DMatlab_ROOT_DIR=/Applications/MATLAB_R2025a.app
```

to help cmake identify the correct compiler and Matlab.

Upon `build`, all source files compile correctly but linking fails at the final stage:

```
ld: symbol(s) not found for architecture arm64
clang: error: linker command failed with exit code 1 (use -v to see invocation)
gmake[2]: *** [mex/CMakeFiles/mexIgraphDispatcher.dir/build.make:432: mex/mexIgraphDispatcher.mexmaca64] Error 1
```

Here's the full output:

```
[ 96%] Building C object mex/CMakeFiles/mexIgraphDispatcher.dir/Applications/MATLAB_R2025a.app/extern/version/c_mexapi_version.c.o
[ 96%] Linking CXX shared module mexIgraphDispatcher.mexmaca64
Undefined symbols for architecture arm64:
  "std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>>::compare(unsigned long, unsigned long, char const*, unsigned long) const", referenced from:
      bool std::__1::operator==[abi:ne200100]<char, std::__1::char_traits<char>, std::__1::allocator<char>>(std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>> const&, char const*) in libigraph.a(prpack_solver.cpp.o)
  "std::exception::what() const", referenced from:
      vtable for std::__1::bad_function_call in libigraph.a(graph.cc.o)
  "std::logic_error::logic_error(char const*)", referenced from:
      std::length_error::length_error[abi:ne200100](char const*) in libigraph.a(infomap.cc.o)
      std::length_error::length_error[abi:ne200100](char const*) in libigraph.a(infomap_Greedy.cc.o)
      std::length_error::length_error[abi:ne200100](char const*) in libigraph.a(infomap_FlowGraph.cc.o)
      std::length_error::length_error[abi:ne200100](char const*) in libigraph.a(degree_sequence.cpp.o)
      std::length_error::length_error[abi:ne200100](char const*) in libigraph.a(graph.cc.o)
      std::length_error::length_error[abi:ne200100](char const*) in libigraph.a(prpack_base_graph.cpp.o)
      std::invalid_argument::invalid_argument[abi:ne200100](char const*) in libigraph.a(gengraph_graph_molloy_hash.cpp.o)
      ...
  "std::logic_error::logic_error(std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>> const&)", referenced from:
      std::invalid_argument::invalid_argument[abi:ne200100](std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>> const&) in libigraph.a(prpack_solver.cpp.o)
  "std::length_error::~length_error()", referenced from:
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(infomap.cc.o)
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(infomap_Greedy.cc.o)
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(infomap_FlowGraph.cc.o)
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(degree_sequence.cpp.o)
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(graph.cc.o)
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(prpack_base_graph.cpp.o)
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(utils.cc.o)
      ...
  "std::runtime_error::runtime_error(char const*)", referenced from:
      (anonymous namespace)::bliss_from_igraph(igraph_s const*) in libigraph.a(bliss.cc.o)
      bliss::Digraph::add_edge(unsigned int, unsigned int) in libigraph.a(graph.cc.o)
      bliss::Graph::add_edge(unsigned int, unsigned int) in libigraph.a(graph.cc.o)
      prpack::prpack_base_graph::prpack_base_graph(prpack::prpack_csr const*) in libigraph.a(prpack_base_graph.cpp.o)
      gengraph::graph_molloy_opt::breadth_path_search(long long, long long*, double*, unsigned char*) in libigraph.a(gengraph_graph_molloy_optimized.cpp.o)
  "std::runtime_error::~runtime_error()", referenced from:
      (anonymous namespace)::bliss_from_igraph(igraph_s const*) in libigraph.a(bliss.cc.o)
      bliss::Digraph::add_edge(unsigned int, unsigned int) in libigraph.a(graph.cc.o)
      bliss::Graph::add_edge(unsigned int, unsigned int) in libigraph.a(graph.cc.o)
      prpack::prpack_base_graph::prpack_base_graph(prpack::prpack_csr const*) in libigraph.a(prpack_base_graph.cpp.o)
      gengraph::graph_molloy_opt::breadth_path_search(long long, long long*, double*, unsigned char*) in libigraph.a(gengraph_graph_molloy_optimized.cpp.o)
  "std::invalid_argument::~invalid_argument()", referenced from:
      prpack::prpack_solver::solve(double, double, double const*, double const*, char const*) in libigraph.a(prpack_solver.cpp.o)
      gengraph::graph_molloy_hash::shuffle(long long, long long, int) in libigraph.a(gengraph_graph_molloy_hash.cpp.o)
  "std::bad_array_new_length::bad_array_new_length()", referenced from:
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(infomap.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(infomap_Greedy.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(infomap_FlowGraph.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(bliss.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(degree_sequence.cpp.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(graph.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(prpack_base_graph.cpp.o)
      ...
  "std::bad_array_new_length::~bad_array_new_length()", referenced from:
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(infomap.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(infomap_Greedy.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(infomap_FlowGraph.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(bliss.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(degree_sequence.cpp.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(graph.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(prpack_base_graph.cpp.o)
      ...
  "std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>>::__init(char const*, unsigned long)", referenced from:
      std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>>::basic_string[abi:ne200100]<0>(char const*) in libigraph.a(prpack_solver.cpp.o)
  "std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>>::append(char const*)", referenced from:
      std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>>::operator+=[abi:ne200100](char const*) in libigraph.a(prpack_solver.cpp.o)
      std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>> std::__1::operator+[abi:ne200100]<char, std::__1::char_traits<char>, std::__1::allocator<char>>(std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>>&&, char const*) in libigraph.a(prpack_solver.cpp.o)
  "std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>>::assign(char const*)", referenced from:
      std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>>::operator=[abi:ne200100](char const*) in libigraph.a(prpack_solver.cpp.o)
  "std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>>::~basic_string()", referenced from:
      prpack::prpack_result::~prpack_result() in libigraph.a(prpack_result.cpp.o)
      prpack::prpack_solver::solve(double, double, double const*, double const*, char const*) in libigraph.a(prpack_solver.cpp.o)
  "std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>>::operator=(std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>> const&)", referenced from:
      prpack::prpack_solver::solve(double, double, double const*, double const*, char const*) in libigraph.a(prpack_solver.cpp.o)
  "void std::__1::__sort<std::__1::__less<unsigned int, unsigned int>&, unsigned int*>(unsigned int*, unsigned int*, std::__1::__less<unsigned int, unsigned int>&)", referenced from:
      void std::__1::__sort_dispatch[abi:ne200100]<std::__1::_ClassicAlgPolicy, unsigned int, 0>(unsigned int*, unsigned int*, std::__1::__less<void, void>&) in libigraph.a(graph.cc.o)
  "std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>> std::__1::operator+<char, std::__1::char_traits<char>, std::__1::allocator<char>>(char const*, std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>> const&)", referenced from:
      prpack::prpack_solver::solve(double, double, double const*, double const*, char const*) in libigraph.a(prpack_solver.cpp.o)
  "std::bad_alloc::bad_alloc()", referenced from:
      (anonymous namespace)::AutCollector::operator()(unsigned int, unsigned int const*) in libigraph.a(bliss.cc.o)
  "std::bad_alloc::~bad_alloc()", referenced from:
      (anonymous namespace)::AutCollector::operator()(unsigned int, unsigned int const*) in libigraph.a(bliss.cc.o)
  "std::exception::~exception()", referenced from:
      std::__1::bad_function_call::~bad_function_call() in libigraph.a(graph.cc.o)
  "std::nothrow", referenced from:
      std::__1::unique_ptr<vbd_pair, std::__1::__temporary_buffer_deleter<vbd_pair>> std::__1::__allocate_unique_temporary_buffer[abi:ne200100]<vbd_pair>(long) in libigraph.a(degree_sequence.cpp.o)
      std::__1::unique_ptr<vd_pair, std::__1::__temporary_buffer_deleter<vd_pair>> std::__1::__allocate_unique_temporary_buffer[abi:ne200100]<vd_pair>(long) in libigraph.a(degree_sequence.cpp.o)
  "std::terminate()", referenced from:
      ___clang_call_terminate in libigraph.a(infomap.cc.o)
      ___clang_call_terminate in libigraph.a(clustertool.cpp.o)
      ___clang_call_terminate in libigraph.a(walktrap.cpp.o)
      ___clang_call_terminate in libigraph.a(pottsmodel_2.cpp.o)
      ___clang_call_terminate in libigraph.a(infomap_Greedy.cc.o)
      ___clang_call_terminate in libigraph.a(infomap_FlowGraph.cc.o)
      ___clang_call_terminate in libigraph.a(gengraph_mr-connected.cpp.o)
      ...
  "typeinfo for std::range_error", referenced from:
      GCC_except_table0 in libigraph.a(infomap.cc.o)
      GCC_except_table0 in libigraph.a(clustertool.cpp.o)
      GCC_except_table4 in libigraph.a(clustertool.cpp.o)
      GCC_except_table0 in libigraph.a(walktrap.cpp.o)
      GCC_except_table13 in libigraph.a(gengraph_mr-connected.cpp.o)
      GCC_except_table0 in libigraph.a(prpack.cpp.o)
      GCC_except_table0 in libigraph.a(bliss.cc.o)
      ...
  "typeinfo for std::length_error", referenced from:
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(infomap.cc.o)
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(infomap_Greedy.cc.o)
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(infomap_FlowGraph.cc.o)
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(degree_sequence.cpp.o)
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(graph.cc.o)
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(prpack_base_graph.cpp.o)
      std::__1::__throw_length_error[abi:ne200100](char const*) in libigraph.a(utils.cc.o)
      ...
  "typeinfo for std::runtime_error", referenced from:
      (anonymous namespace)::bliss_from_igraph(igraph_s const*) in libigraph.a(bliss.cc.o)
      bliss::Digraph::add_edge(unsigned int, unsigned int) in libigraph.a(graph.cc.o)
      bliss::Graph::add_edge(unsigned int, unsigned int) in libigraph.a(graph.cc.o)
      prpack::prpack_base_graph::prpack_base_graph(prpack::prpack_csr const*) in libigraph.a(prpack_base_graph.cpp.o)
      gengraph::graph_molloy_opt::breadth_path_search(long long, long long*, double*, unsigned char*) in libigraph.a(gengraph_graph_molloy_optimized.cpp.o)
  "typeinfo for std::invalid_argument", referenced from:
      prpack::prpack_solver::solve(double, double, double const*, double const*, char const*) in libigraph.a(prpack_solver.cpp.o)
      gengraph::graph_molloy_hash::shuffle(long long, long long, int) in libigraph.a(gengraph_graph_molloy_hash.cpp.o)
  "typeinfo for std::bad_array_new_length", referenced from:
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(infomap.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(infomap_Greedy.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(infomap_FlowGraph.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(bliss.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(degree_sequence.cpp.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(graph.cc.o)
      std::__throw_bad_array_new_length[abi:ne200100]() in libigraph.a(prpack_base_graph.cpp.o)
      ...
  "typeinfo for std::bad_alloc", referenced from:
      GCC_except_table0 in libigraph.a(infomap.cc.o)
      GCC_except_table0 in libigraph.a(clustertool.cpp.o)
      GCC_except_table4 in libigraph.a(clustertool.cpp.o)
      GCC_except_table0 in libigraph.a(walktrap.cpp.o)
      GCC_except_table13 in libigraph.a(gengraph_mr-connected.cpp.o)
      GCC_except_table0 in libigraph.a(prpack.cpp.o)
      (anonymous namespace)::AutCollector::operator()(unsigned int, unsigned int const*) in libigraph.a(bliss.cc.o)
      ...
  "typeinfo for std::exception", referenced from:
      GCC_except_table0 in libigraph.a(infomap.cc.o)
      GCC_except_table0 in libigraph.a(clustertool.cpp.o)
      GCC_except_table4 in libigraph.a(clustertool.cpp.o)
      GCC_except_table0 in libigraph.a(walktrap.cpp.o)
      GCC_except_table13 in libigraph.a(gengraph_mr-connected.cpp.o)
      GCC_except_table0 in libigraph.a(prpack.cpp.o)
      GCC_except_table0 in libigraph.a(bliss.cc.o)
      ...
  "vtable for __cxxabiv1::__class_type_info", referenced from:
      typeinfo for DLList<NNode*> in libigraph.a(clustertool.cpp.o)
      typeinfo for DLList<NLink*> in libigraph.a(clustertool.cpp.o)
      typeinfo for DLList<DLList<NNode*>*> in libigraph.a(clustertool.cpp.o)
      typeinfo for DLList<NNode*> in libigraph.a(NetRoutines.cpp.o)
      typeinfo for DLList<NLink*> in libigraph.a(NetRoutines.cpp.o)
      typeinfo for DLList<long long*> in libigraph.a(pottsmodel_2.cpp.o)
      typeinfo for DLList<NNode*> in libigraph.a(pottsmodel_2.cpp.o)
      ...
  NOTE: a missing vtable usually means the first non-inline virtual member function has no definition.
  "vtable for __cxxabiv1::__si_class_type_info", referenced from:
      typeinfo for DL_Indexed_List<NNode*> in libigraph.a(clustertool.cpp.o)
      typeinfo for DL_Indexed_List<NLink*> in libigraph.a(clustertool.cpp.o)
      typeinfo for DL_Indexed_List<DLList<NNode*>*> in libigraph.a(clustertool.cpp.o)
      typeinfo for DL_Indexed_List<long long*> in libigraph.a(pottsmodel_2.cpp.o)
      typeinfo for std::__1::__function::__func<(anonymous namespace)::AbortChecker, std::__1::allocator<(anonymous namespace)::AbortChecker>, bool ()> in libigraph.a(bliss.cc.o)
      typeinfo for std::__1::__function::__func<(anonymous namespace)::AutCollector, std::__1::allocator<(anonymous namespace)::AutCollector>, void (unsigned int, unsigned int const*)> in libigraph.a(bliss.cc.o)
      typeinfo for bliss::Digraph in libigraph.a(graph.cc.o)
      ...
  NOTE: a missing vtable usually means the first non-inline virtual member function has no definition.
  "vtable for std::length_error", referenced from:
      std::length_error::length_error[abi:ne200100](char const*) in libigraph.a(infomap.cc.o)
      std::length_error::length_error[abi:ne200100](char const*) in libigraph.a(infomap_Greedy.cc.o)
      std::length_error::length_error[abi:ne200100](char const*) in libigraph.a(infomap_FlowGraph.cc.o)
      std::length_error::length_error[abi:ne200100](char const*) in libigraph.a(degree_sequence.cpp.o)
      std::length_error::length_error[abi:ne200100](char const*) in libigraph.a(graph.cc.o)
      std::length_error::length_error[abi:ne200100](char const*) in libigraph.a(prpack_base_graph.cpp.o)
      std::length_error::length_error[abi:ne200100](char const*) in libigraph.a(utils.cc.o)
      ...
  NOTE: a missing vtable usually means the first non-inline virtual member function has no definition.
  "vtable for std::invalid_argument", referenced from:
      std::invalid_argument::invalid_argument[abi:ne200100](std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>> const&) in libigraph.a(prpack_solver.cpp.o)
      std::invalid_argument::invalid_argument[abi:ne200100](char const*) in libigraph.a(gengraph_graph_molloy_hash.cpp.o)
  NOTE: a missing vtable usually means the first non-inline virtual member function has no definition.
  "vtable for std::exception", referenced from:
      std::exception::exception[abi:ne200100]() in libigraph.a(graph.cc.o)
  NOTE: a missing vtable usually means the first non-inline virtual member function has no definition.
  "operator delete[](void*)", referenced from:
      HugeArray<DLItem<DLList<NNode*>*>*>::~HugeArray() in libigraph.a(clustertool.cpp.o)
      HugeArray<DLItem<NLink*>*>::~HugeArray() in libigraph.a(clustertool.cpp.o)
      HugeArray<DLItem<NNode*>*>::~HugeArray() in libigraph.a(clustertool.cpp.o)
      PottsModel::~PottsModel() in libigraph.a(pottsmodel_2.cpp.o)
      PottsModelN::~PottsModelN() in libigraph.a(pottsmodel_2.cpp.o)
      PottsModelN::assign_initial_conf(bool) in libigraph.a(pottsmodel_2.cpp.o)
      PottsModelN::WriteClusters(double*, double*, igraph_vector_int_t*, igraph_vector_int_t*, igraph_matrix_t*, igraph_matrix_t*, double*, double, double, double) in libigraph.a(pottsmodel_2.cpp.o)
      ...
  "operator delete(void*)", referenced from:
      void std::__1::__libcpp_operator_delete[abi:ne200100]<long long*>(long long*) in libigraph.a(infomap.cc.o)
      void std::__1::__libcpp_operator_delete[abi:ne200100]<unsigned long*>(unsigned long*) in libigraph.a(infomap.cc.o)
      void std::__1::__libcpp_operator_delete[abi:ne200100]<double*>(double*) in libigraph.a(infomap.cc.o)
      void std::__1::__libcpp_operator_delete[abi:ne200100]<std::__1::pair<long long, double>*>(std::__1::pair<long long, double>*) in libigraph.a(infomap.cc.o)
      void std::__1::__libcpp_operator_delete[abi:ne200100]<Node*>(Node*) in libigraph.a(infomap.cc.o)
      DLList<NNode*>::DLList() in libigraph.a(clustertool.cpp.o)
      DLList<NNode*>::~DLList() in libigraph.a(clustertool.cpp.o)
      ...
  "operator new[](unsigned long)", referenced from:
      HugeArray<DLItem<NNode*>*>::HugeArray() in libigraph.a(clustertool.cpp.o)
      HugeArray<DLItem<NNode*>*>::Set(long long) in libigraph.a(clustertool.cpp.o)
      HugeArray<DLItem<NLink*>*>::HugeArray() in libigraph.a(clustertool.cpp.o)
      HugeArray<DLItem<NLink*>*>::Set(long long) in libigraph.a(clustertool.cpp.o)
      HugeArray<DLItem<DLList<NNode*>*>*>::HugeArray() in libigraph.a(clustertool.cpp.o)
      HugeArray<DLItem<DLList<NNode*>*>*>::Set(long long) in libigraph.a(clustertool.cpp.o)
      HugeArray<DLItem<NNode*>*>::Set(long long) in libigraph.a(NetRoutines.cpp.o)
      ...
  "operator new(unsigned long)", referenced from:
      void* std::__1::__libcpp_operator_new[abi:ne200100]<unsigned long>(unsigned long) in libigraph.a(infomap.cc.o)
      DLList<NNode*>::DLList() in libigraph.a(clustertool.cpp.o)
      DL_Indexed_List<NNode*>::pInsert(NNode*, DLItem<NNode*>*) in libigraph.a(clustertool.cpp.o)
[…]
