### 2025-07-22 DavidRConnell/matlab-igraph#5 @platipodium
Downloaded  pre-compiled and installed toolbox fails to run on M1/arm64/Matlab2025
Related to https://github.com/openjournals/joss-reviews/issues/8622#issuecomment-3094739199

Installed toolbox fails to run with missing internal libraries/files.

I could successfully download the toolbox and install it into my Matlab2025, where it shows up in the Add-ons manager.

But, in the test live script, it fails at initialization:

```
graph = igraph.randgame("BarabasiBag", nNodes = 15)

Invalid MEX-file '/Users/Lemmen/Library/Application Support/MathWorks/MATLAB Add-Ons/Toolboxes/matlab-igraph/+igraph/private/mexIgraphDispatcher.mexmaca64': dlopen(/Users/Lemmen/Library/Application Support/MathWorks/MATLAB Add-Ons/Toolboxes/matlab-igraph/+igraph/private/mexIgraphDispatcher.mexmaca64, 0x0006): Library not loaded: @rpath/libMatlabEngine.dylib
  Referenced from: <C147BFE2-8D11-3B27-8C37-8BF1080FAEE0> /Users/Lemmen/Library/Application Support/MathWorks/MATLAB Add-Ons/Toolboxes/matlab-igraph/+igraph/private/mexIgraphDispatcher.mexmaca64
  Reason: tried: '/Applications/MATLAB_R2025a.app/bin/maca64/libMatlabEngine.dylib' (no such file), '/Applications/MATLAB_R2025a.app/bin/maca64/./libMatlabEngine.dylib' (no such file), '/Applications/MATLAB_R2025a.app/bin/maca64/../../sys/os/maca64/libMatlabEngine.dylib' (no such file), '/Applications/MATLAB_R2025a.app/bin/maca64/libMatlabEngine.dylib' (no such file), '/Applications/MATLAB_R2025a.app/bin/maca64/./libMatlabEngine.dylib' (no such file), '/Applications/MATLAB_R2025a.app/bin/maca64/../../sys/os/maca64/libMatlabEngine.dylib' (no such file), '/Applications/MATLAB_R2025a.app/bin/maca64/../../runtime/maca64/libMatlabEngine.dylib' (no such file), '/Applications/MATLAB_R2025a.app/bin/maca64/../../sys/java/jre/maca64/jre/lib/server/libMatlabEngine.dylib' (no such file), '/Applications/MATLAB_R2025a.app/bin/maca64/../../sys/java/jre/maca64/jre/lib/libMatlabEngine.dylib' (no such file)

Error in [igraph.randgame](matlab:matlab.lang.internal.introspective.errorDocCallback('igraph.randgame', '/Users/Lemmen/Library/Application Support/MathWorks/MATLAB Add-Ons/Toolboxes/matlab-igraph/+igraph/randgame.m', 633)) ([line 633](matlab: opentoline('/Users/Lemmen/Library/Application Support/MathWorks/MATLAB Add-Ons/Toolboxes/matlab-igraph/+igraph/randgame.m',633,0)))
    graph = mexIgraphDispatcher(mfilename(), method, graphOpts, methodOpts);
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

```
