### 2025-07-22 DavidRConnell/matlab-igraph#5 @platipodium
It also fails with the same error on my M2/Matlab2024
