### 2025-07-22 DavidRConnell/matlab-igraph#5 @DavidRConnell
Thanks for bringing this to my attention. I don't have a macOS device to work with but I am running tests on GitHub actions for macOS. I have just ran a double check test ([results here](https://github.com/DavidRConnell/matlab-igraph/actions/runs/16452801980/job/46502341683)) and it is running correctly on actions in the arm and intel macOS containers (with Matlab 2024). (And that includes a check that the same Barabasi bag generator that failed on your computer runs.)

I imagine I might be assuming a non-standard path or install location that is in the actions containers is standard on everyone's computer when it's not? Specifically, you mentioned in [#4] you had to explicitly pass the path to matlab. The matlab setup code is possibly setting some important values that are not default. At the least it brings up that the linker can't find the correct paths so the current precombiled binaries appear to be fragile to the exact matlab install.

I will see what I can do to reproduce this and try to get a more flexible build. Unfortunately, I do not have too much time to spend on this at the moment with my defense coming up (unrelated) so I thank you for taking the time and apologize if I don't end up finding a fix soon and have to post-pone my submission. I think I was idealistic about how smoothly this was going to go.
