### 2025-07-23 3945795
Turn off implicit linking to MATLAB libraries
Fixes #5

On macOS a MATLBA bug is causing the engine library to not be found.
