+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Status handler no longer adds a newline character to the message, that is left to the caller to do.
 - Misspelling of fatal.
 - Prevent collecting a duplicate edge when converting symmetric sparse matrices to igraph graphs.
+- Running on macOS. Was failing due to dynamic linker unable to find the matlab engine library. No longer link to that library since it isn't needed anyway.
 
 ## [0.2.0] 2024-09-03
 
