### 2025-07-23 DavidRConnell/matlab-igraph#5 @DavidRConnell
Thank you. I had actually been hoping someone on an arm macs would try it for this very reason.

I think I am just figuring out the issue. If you wouldn't mind could you try [this toolbox](https://github.com/DavidRConnell/matlab-igraph/actions/runs/16465103657/artifacts/3594743394) (built in this [action](https://github.com/DavidRConnell/matlab-igraph/actions/runs/16465103657))? I don't want to publish a release until it is working.

I was able to reproduce the error by building in one container then testing in a second container. It looks like the error is due to a known bug (although Mathworks says it only effects intel macOS and should be fixed now). Regardless, for some reason the linker is looking in the wrong place for the engine library. But CMake defaults to linking against this. So turned off linking against the engine library and it appears to be fixed.

Annoyingly, now the Windows tests are failing, but I'll take a look at this tomorrow.
