### 2025-07-23 DavidRConnell/matlab-igraph#5 @DavidRConnell
Great. I'll try to send a new release to file-exchange soon.
