### 2025-07-23 8b44925
Remove installation of tools already present in the container
