+++ .github/workflows/build.yaml
@@ -136,7 +136,7 @@ jobs:
           fetch-depth: 0
 
       - name: Install dependencies
-        run: brew install ninja autoconf automake libtool
+        run: brew install autoconf automake libtool
 
       - name: Set up MATLAB
         uses: matlab-actions/setup-matlab@v2
@@ -185,7 +185,7 @@ jobs:
           fetch-depth: 0
 
       - name: Install dependencies
-        run: brew install ninja autoconf automake libtool
+        run: brew install automake
 
       - name: Set up MATLAB
         uses: matlab-actions/setup-matlab@v2
