### 2025-07-23 b82eee7
Perform tests in new containers after installing toolbox
- We had been running the tests on the build/ directory generated by CMake.
Sometimes files that run in the build/ directory don't run after installation
due to differences in which dependencies get linked (or if the the dependencies
weren't shipped correctly leading to linker not finding them at all).
- This changes testing by installing the toolbox in fresh containers and making
sure the tests run there.
