+++ .github/workflows/release.yaml
@@ -31,39 +31,41 @@ jobs:
     needs: publish-docs
     uses: ./.github/workflows/hugo.yaml
 
-  release:
-    name: Release new packages
-    runs-on: ubuntu-latest
-    if: ${{ github.ref_type == 'tag' }}
-    needs: build-toolbox
-    env:
-      TAG_NAME: ${{ github.ref_name }}
+  # Automated releases fail on File-Exchange side since they don't wait for
+  # GitHub to upload the files before creating the File-Exchange release.
+  # release:
+  #   name: Release new packages
+  #   runs-on: ubuntu-latest
+  #   if: ${{ github.ref_type == 'tag' }}
+  #   needs: build-toolbox
+  #   env:
+  #     TAG_NAME: ${{ github.ref_name }}
 
-    steps:
-      - name: Create release name
-        run: echo RELEASE_NAME=${TAG_NAME/v/matlab-igraph } >> $GITHUB_ENV
+  #   steps:
+  #     - name: Create release name
+  #       run: echo RELEASE_NAME=${TAG_NAME/v/matlab-igraph } >> $GITHUB_ENV
 
-      - uses: actions/checkout@v4
-        with:
-          submodules: true
-          fetch-depth: 0
+  #     - uses: actions/checkout@v4
+  #       with:
+  #         submodules: true
+  #         fetch-depth: 0
 
-      - name: Create release notes
-        run: |
-          range=$(grep --max-count=2 --line-number "## \[" < CHANGELOG.md |\
-            cut -f 1 -d: | tr '\n' ',' | sed -n 's/,$/p\n/p')
-          sed -n $range < CHANGELOG.md | head -n-2 | tail -n+3 > release_notes.md
-          sed -i 's/###/##/g' release_notes.md
+  #     - name: Create release notes
+  #       run: |
+  #         range=$(grep --max-count=2 --line-number "## \[" < CHANGELOG.md |\
+  #           cut -f 1 -d: | tr '\n' ',' | sed -n 's/,$/p\n/p')
+  #         sed -n $range < CHANGELOG.md | head -n-2 | tail -n+3 > release_notes.md
+  #         sed -i 's/###/##/g' release_notes.md
 
-      - name: Collect package
-        uses: actions/download-artifact@v4
-        with:
-          name: matlab-igraph-toolbox-${{ github.sha }}
-          path: release
+  #     - name: Collect package
+  #       uses: actions/download-artifact@v4
+  #       with:
+  #         name: matlab-igraph-toolbox-${{ github.sha }}
+  #         path: release
 
-      - name: Upload assets to release
-        uses: ncipollo/release-action@v1
-        with:
-          name: ${{ env.RELEASE_NAME }}
-          bodyFile: release_notes.md
-          artifacts: release/*
+  #     - name: Upload assets to release
+  #       uses: ncipollo/release-action@v1
+  #       with:
+  #         name: ${{ env.RELEASE_NAME }}
+  #         bodyFile: release_notes.md
+  #         artifacts: release/*
