### 2025-07-23 fee5703
Fix build to ignore non-toolbox artifacts
