+++ .github/workflows/build.yaml
@@ -193,7 +193,7 @@ jobs:
         uses: matlab-actions/setup-matlab@v2
 
       - name: Make toolboxes
-        uses: matlab-actions/run-command@v1
+        uses: matlab-actions/run-command@v2
         with:
           command: addpath(".github/workflows"), buildAllToolboxes("artifacts", "release", "${{ github.ref_name }}")
 
