### 2025-07-24 DavidRConnell/matlab-igraph#4 @platipodium
Note that I can build and link the current and vendored versions of igraph on my system without problems, so this seems not an upstream issue.
