### 2025-07-26 DavidRConnell/matlab-igraph#4 @DavidRConnell
When you get a chance would you mind deleting build and rerunning the initial config cmake command and printing the detection info?^

Also when you built igraph, was that creating a shared lib or a static lib?

I am trying to figure out what might be going wrong here and it seems like it's either 1. Some dependency is not getting compiled 2. Some dependency is getting compiled for the wrong architecture 3. There's an issue with how igraph is getting compiled.

2 seems unlikely given that it is trying to compile mexamaca64 extension so it's not a mistake with the mex architecture logic and I don't see why it would compile something for x86_64. 3. Given that the linking errors start with function calls within igraph, it seems like it could be an error there. So the fact you were able to compile the vendored igraph on its own is important information given that all the CMake build logic for igraph is deferred to igraph's own CMakeLists.txt.

So my guess for now is something is not getting compiled. Maybe for some reason the igraph dependencies are not getting compiled when the static igraph lib is getting compiled?

^ as in this information

``` text
-- The C compiler identification is AppleClang 15.0.0.15000309
-- The CXX compiler identification is AppleClang 15.0.0.15000309
-- Detecting C compiler ABI info
-- Detecting C compiler ABI info - done
-- Check for working C compiler: /Applications/Xcode_15.4.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/cc - skipped
-- Detecting C compile features
-- Detecting C compile features - done
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /Applications/Xcode_15.4.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/c++ - skipped
-- Detecting CXX compile features
-- Found Matlab: /Users/runner/hostedtoolcache/MATLAB/2025.1.999/arm64/MATLAB.app (found version "25.1.0.2973910") found components: MAIN_PROGRAM MAT_LIBRARY MX_LIBRARY MEX_COMPILER
-- Matlab: 25.1.0.2973910
-- Version number from Git: 0.10.13
-- Performing Test COMPILER_SUPPORTS_UNKNOWN_WARNING_OPTION_FLAG
-- Performing Test COMPILER_SUPPORTS_UNKNOWN_WARNING_OPTION_FLAG - Success
-- Performing Test COMPILER_HAS_DEPRECATED_ENUMVAL_ATTR
-- Performing Test COMPILER_HAS_DEPRECATED_ENUMVAL_ATTR - Success
-- Thread-local storage: supported (__thread)
-- Performing Test CMAKE_HAVE_LIBC_PTHREAD
-- Performing Test CMAKE_HAVE_LIBC_PTHREAD - Success
-- Found Threads: TRUE
-- Found FLEX: /usr/bin/flex (found version "2.6.4")
-- Found BISON: /usr/bin/bison (found version "2.3")
-- Looking for strcasecmp
-- Looking for strcasecmp - found
-- Looking for strncasecmp
-- Looking for strncasecmp - found
-- Looking for _stricmp
-- Looking for _stricmp - not found
-- Looking for _strnicmp
-- Looking for _strnicmp - not found
-- Looking for strdup
-- Looking for strdup - found
-- Looking for strndup
-- Looking for strndup - found
-- Looking for include file xlocale.h
-- Looking for include file xlocale.h - found
-- Looking for uselocale
-- Looking for uselocale - found
-- Looking for _configthreadlocale
-- Looking for _configthreadlocale - not found
-- Performing Test HAVE__UMUL128
-- Performing Test HAVE__UMUL128 - Failed
-- Performing Test HAVE___UMULH
-- Performing Test HAVE___UMULH - Failed
-- Looking for C++ include sys/types.h
-- Looking for C++ include sys/types.h - found
-- Looking for C++ include stdint.h
-- Looking for C++ include stdint.h - found
-- Looking for C++ include stddef.h
-- Looking for C++ include stddef.h - found
-- Check size of __uint128_t
-- Check size of __uint128_t - done
-- Performing Test HAVE__POPCNT64
-- Performing Test HAVE__POPCNT64 - Failed
-- Performing Test HAVE__POPCNT
-- Performing Test HAVE__POPCNT - Failed
-- Performing Test HAVE__BITSCANFORWARD64
-- Performing Test HAVE__BITSCANFORWARD64 - Failed
-- Performing Test HAVE__BITSCANFORWARD
-- Performing Test HAVE__BITSCANFORWARD - Failed
-- Performing Test HAVE__BITSCANREVERSE64
-- Performing Test HAVE__BITSCANREVERSE64 - Failed
-- Performing Test HAVE__BITSCANREVERSE
-- Performing Test HAVE__BITSCANREVERSE - Failed
-- Performing Test HAVE_BUILTIN_OVERFLOW
-- Performing Test HAVE_BUILTIN_OVERFLOW - Success
-- Looking for _LIBCPP_VERSION
-- Looking for _LIBCPP_VERSION - found
-- Looking for __GLIBCXX__
-- Looking for __GLIBCXX__ - not found
-- Performing Test COMPILER_HAS_HIDDEN_VISIBILITY
-- Performing Test COMPILER_HAS_HIDDEN_VISIBILITY - Success
-- Performing Test COMPILER_HAS_HIDDEN_INLINE_VISIBILITY
-- Performing Test COMPILER_HAS_HIDDEN_INLINE_VISIBILITY - Success
-- Performing Test COMPILER_HAS_DEPRECATED_ATTR
-- Performing Test COMPILER_HAS_DEPRECATED_ATTR - Success
--
-- -----[ Build configuration ]----
-- Version:                 0.10.13
-- CMake build type:        default
-- Library type:             static
-- igraph_integer_t size:   64 bits
--
-- ----------[ Features ]----------
-- GLPK for optimization:       yes
-- Reading GraphML files:       yes
-- Thread-local storage:        yes
-- Link-time optimization:       no
--
-- --------[ Dependencies ]--------
-- ARPACK:                 vendored
-- BISON:                       yes
-- BLAS:                   vendored
-- FLEX:                        yes
-- GLPK:                   vendored
-- GMP:                    vendored
-- LAPACK:                 vendored
-- LibXml2:                     yes
-- OpenMP:                       no
-- PLFIT:                  vendored
--
-- -----------[ Testing ]----------
-- Diff tool:             not found
-- Sanitizers:                 none
-- Code coverage:                no
-- Verify 'finally' stack:       no
--
-- --------[ Documentation ]-------
-- HTML:                         no
-- PDF:                          no
-- INFO:                         no
--
-- igraph configured successfully.
--
-- matlab-igraph version: v0.2.1
-- MEX extension: mexmaca64
-- Matlab arch: maca64
```
