### 2025-07-26 DavidRConnell/matlab-igraph#4 @platipodium
Sure, here's my cmake dump.  To me it seems correct but there are several warnings about shadowed libraries at the end.

```
08:24:56 ✖  cmake .. -DCMAKE_C_COMPILER=${CONDA_PREFIX}/bin/clang -DCMAKE_CXX_COMPILER=${CONDA_PREFIX}/bin/clang++ -DMatlab_ROOT_DIR=/Applications/MATLAB_R2025a.app
-- Found Matlab: /Applications/MATLAB_R2025a.app (found version "25.1.0.2943329") found components: MAIN_PROGRAM MAT_LIBRARY MX_LIBRARY MEX_COMPILER
-- Matlab: 25.1.0.2943329
-- Version number from Git: 0.10.13
-- Performing Test COMPILER_SUPPORTS_UNKNOWN_WARNING_OPTION_FLAG
-- Performing Test COMPILER_SUPPORTS_UNKNOWN_WARNING_OPTION_FLAG - Success
-- Performing Test COMPILER_HAS_DEPRECATED_ENUMVAL_ATTR
-- Performing Test COMPILER_HAS_DEPRECATED_ENUMVAL_ATTR - Success
-- Thread-local storage: supported (__thread)
-- Performing Test CMAKE_HAVE_LIBC_PTHREAD
-- Performing Test CMAKE_HAVE_LIBC_PTHREAD - Success
-- Found Threads: TRUE
-- Found FLEX: /opt/conda/envs/esmf/bin/flex (found version "2.6.4")
-- Found BISON: /opt/conda/envs/esmf/bin/bison (found version "3.8.2")
-- Looking for strcasecmp
-- Looking for strcasecmp - found
-- Looking for strncasecmp
-- Looking for strncasecmp - found
-- Looking for _stricmp
-- Looking for _stricmp - not found
-- Looking for _strnicmp
-- Looking for _strnicmp - not found
-- Looking for strdup
-- Looking for strdup - found
-- Looking for strndup
-- Looking for strndup - found
-- Looking for include file xlocale.h
-- Looking for include file xlocale.h - found
-- Looking for uselocale
-- Looking for uselocale - found
-- Looking for _configthreadlocale
-- Looking for _configthreadlocale - not found
-- Performing Test HAVE__UMUL128
-- Performing Test HAVE__UMUL128 - Failed
-- Performing Test HAVE___UMULH
-- Performing Test HAVE___UMULH - Failed
-- Looking for C++ include sys/types.h
-- Looking for C++ include sys/types.h - found
-- Looking for C++ include stdint.h
-- Looking for C++ include stdint.h - found
-- Looking for C++ include stddef.h
-- Looking for C++ include stddef.h - found
-- Check size of __uint128_t
-- Check size of __uint128_t - done
-- Performing Test HAVE__POPCNT64
-- Performing Test HAVE__POPCNT64 - Failed
-- Performing Test HAVE__POPCNT
-- Performing Test HAVE__POPCNT - Failed
-- Performing Test HAVE__BITSCANFORWARD64
-- Performing Test HAVE__BITSCANFORWARD64 - Failed
-- Performing Test HAVE__BITSCANFORWARD
-- Performing Test HAVE__BITSCANFORWARD - Failed
-- Performing Test HAVE__BITSCANREVERSE64
-- Performing Test HAVE__BITSCANREVERSE64 - Failed
-- Performing Test HAVE__BITSCANREVERSE
-- Performing Test HAVE__BITSCANREVERSE - Failed
-- Performing Test HAVE_BUILTIN_OVERFLOW
-- Performing Test HAVE_BUILTIN_OVERFLOW - Success
-- Looking for _LIBCPP_VERSION
-- Looking for _LIBCPP_VERSION - found
-- Looking for __GLIBCXX__
-- Looking for __GLIBCXX__ - not found
-- Performing Test COMPILER_HAS_HIDDEN_VISIBILITY
-- Performing Test COMPILER_HAS_HIDDEN_VISIBILITY - Success
-- Performing Test COMPILER_HAS_HIDDEN_INLINE_VISIBILITY
-- Performing Test COMPILER_HAS_HIDDEN_INLINE_VISIBILITY - Success
-- Performing Test COMPILER_HAS_DEPRECATED_ATTR
-- Performing Test COMPILER_HAS_DEPRECATED_ATTR - Success
--
-- -----[ Build configuration ]----
-- Version:                 0.10.13
-- CMake build type:        default
-- Library type:             static
-- igraph_integer_t size:   64 bits
-- Compiler cache:           ccache
--
-- ----------[ Features ]----------
-- GLPK for optimization:       yes
-- Reading GraphML files:       yes
-- Thread-local storage:        yes
-- Link-time optimization:       no
--
-- --------[ Dependencies ]--------
-- ARPACK:                 vendored
-- BISON:                       yes
-- BLAS:                   vendored
-- FLEX:                        yes
-- GLPK:                   vendored
-- GMP:                    vendored
-- LAPACK:                 vendored
-- LibXml2:                     yes
-- OpenMP:                      yes
-- PLFIT:                  vendored
--
-- -----------[ Testing ]----------
-- Diff tool:             not found
-- Sanitizers:                 none
-- Code coverage:                no
-- Verify 'finally' stack:       no
--
-- --------[ Documentation ]-------
-- HTML:                         no
-- PDF:                          no
-- INFO:                         no
--
-- igraph configured successfully.
--
-- matlab-igraph version: v0.2.0
-- MEX extension: mexmaca64
-- Matlab arch: maca64
--
-- Configuring done (18.8s)
CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target mexIgraphDispatcher
  because files in some directories may conflict with libraries in implicit
  directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  mex/CMakeLists.txt:5 (matlab_add_mex)

CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target mexIgraphVersion
  because files in some directories may conflict with libraries in implicit
  directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  mex/CMakeLists.txt:17 (matlab_add_mex)

CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target mexIgraphIsDirected
  because files in some directories may conflict with libraries in implicit
  directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  mex/CMakeLists.txt:17 (matlab_add_mex)

CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target igraphRandi because
  files in some directories may conflict with libraries in implicit
  directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  tests/CMakeLists.txt:10 (matlab_add_mex)

CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target testIsDirected
  because files in some directories may conflict with libraries in implicit
  directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  tests/CMakeLists.txt:10 (matlab_add_mex)

CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target testIsSquare because
  files in some directories may conflict with libraries in implicit
  directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  tests/CMakeLists.txt:10 (matlab_add_mex)

CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target testIsSymmetric
  because files in some directories may conflict with libraries in implicit
  directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  tests/CMakeLists.txt:10 (matlab_add_mex)

CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target testIsTriL because
  files in some directories may conflict with libraries in implicit
  directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  tests/CMakeLists.txt:10 (matlab_add_mex)

CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target testIsTriU because
  files in some directories may conflict with libraries in implicit
  directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  tests/CMakeLists.txt:10 (matlab_add_mex)

CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target testIsVector because
  files in some directories may conflict with libraries in implicit
  directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  tests/CMakeLists.txt:10 (matlab_add_mex)

CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target testIsWeighted
  because files in some directories may conflict with libraries in implicit
  directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  tests/CMakeLists.txt:10 (matlab_add_mex)

CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target testReproduceAdj
  because files in some directories may conflict with libraries in implicit
  directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  tests/CMakeLists.txt:10 (matlab_add_mex)

CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target
  testReproducePartition because files in some directories may conflict with
  libraries in implicit directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  tests/CMakeLists.txt:10 (matlab_add_mex)

CMake Warning at /opt/conda/envs/esmf/share/cmake-4.0/Modules/FindMatlab.cmake:1207 (add_library):
  Cannot generate a safe runtime search path for target convertGraph because
  files in some directories may conflict with libraries in implicit
  directories:

    runtime library [libxml2.dylib] in /opt/conda/envs/esmf/lib may be hidden by files in:
      /Applications/MATLAB_R2025a.app/bin/maca64

  Some of these libraries may not be found correctly.
Call Stack (most recent call first):
  benchmarks/CMakeLists.txt:5 (matlab_add_mex)

-- Generating done (0.5s)
```
