### 2025-07-26 b4c2c31
Test explicitly passing clang++ vs clang as CXX
