+++ .github/workflows/build.yaml
@@ -4,98 +4,7 @@ on:
   - workflow_call
 
 jobs:
-  build-linux:
-    name: Build Linux toolbox
-    runs-on: ubuntu-24.04
-    env:
-      CC: gcc-10
-      CXX: g++-10
-
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          submodules: true
-          fetch-depth: 0
-
-      - name: Install dependencies
-        run: |
-          sudo apt update && sudo apt upgrade
-          sudo apt install cmake ninja-build gcc-10 g++-10 libxml2-dev libbison-dev libfl-dev
-
-      - name: Set up MATLAB
-        uses: matlab-actions/setup-matlab@v2
-
-      - name: Build toolbox
-        run: |
-          cmake -B build . -G Ninja --install-prefix ${GITHUB_WORKSPACE}/dist \
-            -DCMAKE_PACKAGE_VERSION="${{ github.ref_name }}"
-          cmake --build build
-          cmake --install build
-
-      - name: Package dist
-        run: |
-          cmake --build build --target package
-
-      - name: Package source
-        run: cmake --build build --target package_source
-
-      - name: Upload package
-        uses: actions/upload-artifact@v4
-        with:
-          name: matlab-igraph-${{ runner.os }}-${{ github.sha }}
-          path: build/matlab-igraph_*.tar.gz
-
-  build-windows:
-    name: Build Windows toolbox
-    runs-on: windows-latest
-
-    defaults:
-      run:
-        shell: msys2 {0}
-
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          submodules: true
-          fetch-depth: 0
-
-      - name: Install packages
-        uses: msys2/setup-msys2@v2
-        with:
-          msystem: mingw64
-          install: >-
-            git
-            base-devel
-            mingw-w64-x86_64-cmake
-            mingw-w64-x86_64-ninja
-            mingw-w64-x86_64-toolchain
-            mingw-w64-x86_64-libxml2
-          update: true
-          location: D:\
-
-      - name: Set up MATLAB
-        uses: matlab-actions/setup-matlab@v2
-
-      - name: Build toolbox
-        run: |
-          cmake -B build . \
-            -G Ninja \
-            --install-prefix=${GITHUB_WORKSPACE}/dist \
-            -DIGRAPH_OPENMP_SUPPORT=ON \
-            -DCMAKE_PACKAGE_VERSION="${{ github.ref_name }}"
-          cmake --build build --config Release
-          cmake --install build --config Release
-
-      - name: Package dist
-        run: cmake --build build --target package
-
-      - name: Upload package
-        uses: actions/upload-artifact@v4
-        with:
-          name: matlab-igraph-${{ runner.os }}-${{ github.sha }}
-          path: build/matlab-igraph_*.zip
-
-  build-arm-macos:
+  build-macos-with-clangxx:
     name: Build arm macOS toolbox
     runs-on: macos-latest
     env:
@@ -116,6 +25,8 @@ jobs:
       - name: Build toolbox
         run: |
           cmake -B build . -G Ninja \
+            -DCMAKE_C_COMPILER=clang \
+            -DCMAKE_CXX_COMPILER=clang++ \
             --install-prefix=${GITHUB_WORKSPACE}/dist \
             -DCMAKE_PACKAGE_VERSION="${{ github.ref_name }}"
           cmake --build build
@@ -130,9 +41,9 @@ jobs:
           name: matlab-igraph-${{ runner.os }}-arm-${{ github.sha }}
           path: build/matlab-igraph_*.tar.gz
 
-  build-intel-macos:
[…]
