### 2025-07-26 2dc4c57
Test explicitly passing clang++ vs clang as CXX
