### 2025-07-26 DavidRConnell/matlab-igraph#4 @DavidRConnell
I reproduced the issue [in this action](https://github.com/DavidRConnell/matlab-igraph/actions/runs/16541886027). You need to use `clang++` for `CXX` rather than `clang`.
