### 2025-07-26 DavidRConnell/matlab-igraph#4 @platipodium
Hmm,  that is unexpected, as clang++ and clang are actually links to the same compiler executable.

```
21:22:09 ❯ ls -l /opt/conda/envs/esmf/bin/clang++
lrwxr-xr-x 1 Lemmen staff 8 May 30 07:37 /opt/conda/envs/esmf/bin/clang++ -> clang-20
elbe-schism-50627/data/gridded /main 📦 via 🅒 esmf 🐏 6GiB/8GiB | 2GiB/4GiB
21:22:17 ❯ ls -l /opt/conda/envs/esmf/bin/clang
lrwxr-xr-x 1 Lemmen staff 8 May 30 07:37 /opt/conda/envs/esmf/bin/clang -> clang-20
```

But I'll test and report
