### 2025-07-26 DavidRConnell/matlab-igraph#4 @platipodium
Nonetheless, explicitly passing that clang++ succeeds!

`cmake .. -DCMAKE_C_COMPILER=${CONDA_PREFIX}/bin/clang -DCMAKE_CXX_COMPILER=${CONDA_PREFIX}/bin/clang++ -DMatlab_ROOT_DIR=/Applications/MATLAB_R2025a.app`
