### 2025-07-26 DavidRConnell/matlab-igraph#4 @DavidRConnell
On another note, I saw you are on 0.2.0 from the past message. If building yourself, you will need to pull 0.2.1 to fix the other issue or update from the add-on manager.
