### 2025-07-27 #8622 @danielskatz
👋 @gartavanis - can you use the command `@editorialbot generate my checklist` to create your review checklist and get started? @editorialbot commands need to be the first thing in a new comment.
