### 2025-08-19 DavidRConnell/matlab-igraph#6 @platipodium
Clarify copyright
related to https://github.com/openjournals/joss-reviews/issues/8622#issuecomment-3197077134

The copyright statement  indicates that copyright is held by author.  However, in the software paper, the author is affiliated with Rush University Medical College, United States.  Please clarify.
