### 2025-08-19 DavidRConnell/matlab-igraph#7 @platipodium
Ctests fail on arm64/Matlab2025 Symbol not found: ___kmpc_dispatch_deinit
Related to https://github.com/openjournals/joss-reviews/issues/8622#issuecomment-3197077134

I compiled, built and linked successfully on my arm64 M1 using clang and Matlab2025 (see #4 ).  I am now running Ctests and they fail.

```
Test project /Users/Lemmen/devel/review/matlab-igraph/build
    Start 1: TestMxGraph
1/8 Test #1: TestMxGraph ......................   Passed   25.39 sec
    Start 2: TestMxPartition
2/8 Test #2: TestMxPartition ..................   Passed    7.66 sec
    Start 3: TestMxPredicate
3/8 Test #3: TestMxPredicate ..................   Passed    8.64 sec
    Start 4: TestMxRandom
4/8 Test #4: TestMxRandom .....................   Passed    6.18 sec
    Start 5: TestCompare
5/8 Test #5: TestCompare ......................***Failed   14.57 sec
    Start 6: TestGenerators
6/8 Test #6: TestGenerators ...................***Failed   85.12 sec
    Start 7: TestIO
7/8 Test #7: TestIO ...........................***Failed   38.62 sec
    Start 8: TestRNG
8/8 Test #8: TestRNG ..........................***Failed   23.19 sec

50% tests passed, 4 tests failed out of 8

Total Test time (real) = 209.42 sec

The following tests FAILED:
	  5 - TestCompare (Failed)
	  6 - TestGenerators (Failed)
	  7 - TestIO (Failed)
	  8 - TestRNG (Failed)
Errors while running CTest
Output from these tests are in: /Users/Lemmen/devel/review/matlab-igraph/build/Testing/Temporary/LastTest.log
```

[LastTest.log](https://github.com/user-attachments/files/21849643/LastTest.log)
