### 2025-08-19 DavidRConnell/matlab-igraph#8 @platipodium
Clarify usage of "we" in software paper
Related to https://github.com/openjournals/joss-reviews/issues/8622#issuecomment-3197077134

There is only a single author, but the paper repeatedly speaks of "we" as a developer team.  Please clarify.
