### 2025-08-19 DavidRConnell/matlab-igraph#6 @DavidRConnell
Yes, I hold the copyright.
