### 2025-08-19 DavidRConnell/matlab-igraph#8 @danielskatz
I would change this to "I". To me, we implies more than one person.
