### 2025-08-20 DavidRConnell/matlab-igraph#6 @platipodium
Thanks for clarification.  But I still wonder why you would put your affiliation on the paper then ... Closing this as resolved
