### 2025-08-21 90644c5
Change to first person singular
Closes #8.
