### 2025-08-21 d720415
Change to first person singular
Closes #8.
