### 2025-08-21 f429fb2
Change to first person singular
Closes #8.
