### 2025-08-21 b299bb9
Remove affiliation
As mentioned in #6.
