### 2025-08-21 6e911f3
Remove affiliation
As mentioned in #6.
