### 2025-08-21 54b0c10
Remove affiliation
As mentioned in #6.
