### 2025-08-21 DavidRConnell/matlab-igraph#6 @DavidRConnell
I am still rather new to this and didn't consider not including an affiliation. Given that this is my own independent project, I see why the affiliation should not be included so I have removed it.
