### 2025-08-21 DavidRConnell/matlab-igraph#6 @platipodium
@DavidRConnell that is a clean solution.   I've also seen that people  still like to refer to their workplace affiliation, e.g. to demonstrate that they are professionals in the field or to give context to the work.  The way that works is then not to put the affiliation in the title line but make it a footnote with "currently at XXX University", thereby indicating that not the work itself but the person has an affiliation.
