### 2025-08-21 DavidRConnell/matlab-igraph#7 @DavidRConnell
This appears to be related to https://github.com/llvm/llvm-project/issues/98046. If you go below the Linux issue there is repeated issues with macOS that is related to the `libomp` used during compilation. It vaguely looks like it's related to using a different `clang` than the Apple one preinstalled. Someone suggests you may be able to explicitly pass the `openmp` used to `cmake` to the one installed by brew as currently `cmake` is finding a version itself.

If you have tried the pre-compiled toolbox, do you get the error with it? (I am compiling with the standard Apple `clang` compiler.)
