### 2025-08-21 DavidRConnell/matlab-igraph#7 @szhorvat
macOS does not come with `libomp`. If you are distributing binaries, I recommend you either disable OpenMP (which is currently not a big deal for igraph), or build your own `libomp` and link to it statically. The latter solution is a bit complicated and takes time to figure out. I believe we do it for python-igraph, but I did not implement it and do not recall the details.
