### 2025-08-21 DavidRConnell/matlab-igraph#7 @platipodium
Hmm.. in my conda environment, I do have libomp, so I'll try linking to that explicitly.
