### 2025-08-21 DavidRConnell/matlab-igraph#7 @szhorvat
I am not so familiar with how matlab-igraph works, but I very strongly suggest to deactivate conda before you compile _any_ software. Interference from conda environments is the leading cause of compilation failure for R/igraph.  Conda environments don't work well for providing software that you can link to: they are unstable, libraries easily get replaced, and in my experience build systems tend to pick up a mix of libraries from within and outside of the conda environment.

I'm sure you can make it work if you put in the effort, but given that you already see issues:

Just make sure that there are no active conda environments before you compile!!
