### 2025-08-21 DavidRConnell/matlab-igraph#7 @szhorvat
Currently, igraph uses OpenMP in two places: PageRank calculations and fitting power laws. From benchmarking, the speedup is not dramatic for the first one. Furthermore, in some situations (e.g. when parallelizing at a higher level by processing multiple graphs at the same time) having OpenMP on is not desirable. This is subjective, but I think losing OpenMP is not a major disadvantage.

In the future we hope to get much more parallelization, but that is quite challenging due to how igraph does error handling, and our requirement to make as many functions interactively interruptible as possible.

That reminds me to ask: Do you have interruption support working in MATLAB, and if not, do you need help with setting it up?
