### 2025-08-21 DavidRConnell/matlab-igraph#7 @platipodium
> This is of course not part of the review (and you probably don't have time to check), but if this issue also comes up when compiling the igraph C library (https://github.com/igraph/igraph/), do let us know as I'm interested in making this process as easy and as reliable as possible for all users. I suspect we have many academic users who are not experienced in software development, just dabbling in it, and I'm hoping to make this easy for them.
>

Same error for matlab-igraph as for igraph itself with Mac system clang, it's the lack of -fopenmp support

```
...
   clang: error: unsupported option '-fopenmp'
...
```
