### 2025-08-21 cb6c256
Turn off OpenMP support
Closes #7. OpenMP can cause difficulties with compilation particularly on macOS.
At the moment few igraph functions take advantage of it. In the future, may
consider re-enabling it.
