+++ CMakeLists.txt
@@ -26,6 +26,7 @@ set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 set(BUILD_SHARED_LIBS OFF)
 
+set(IGRAPH_OPENMP_SUPPORT OFF)
 set(IGRAPH_USE_INTERNAL_BLAS ON)
 set(IGRAPH_USE_INTERNAL_LAPACK ON)
 set(IGRAPH_USE_INTERNAL_ARPACK ON)
