+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Turn off OpenMP support for now. Prevents issues with macOS users compiling if cmake finds an incompatible copy of OpenMP (#7).
+
 ## [0.2.1] 2025-07-23
 
 ### Add
