### 2025-08-23 22e4110
Bump igraph version to 0.10.16
