### 2025-08-23 fdc7558
Update to igraph 1.0.0
Testing out igraph develop branch ahead of 1.0.0 release (#10).
