+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Bump igraph version to 0.10.16.
+- Bump igraph version to 1.0.0.
 
 ### Fixed
 
