### 2025-08-24 DavidRConnell/matlab-igraph#11 @gartavanis
JOSS Review
Thanks for your work on matlab-igraph and for submitting a paper to JOSS! The paper is well-written and the package well-documented. Below are my comments, they are mostly minor ones:

- [ ] "Development of matlab-igraph18 *begun* with work on the SpeakEasy2 community detection algorith..." (instead of begin)
- [ ] "...a common pattern is *performing* an outer-product like function...": Perhaps "computing" would be a better word choice here.
- [ ] "By using for-loops in C instead of high-level linear algebra packages, the final vector can be created directly, without needing to store a temporary matrix, for large values of 𝑛 and 𝑚 this can be the difference between running out of memory or not.": This sentence is too long, it would be better to break it in two.
- [ ] The paper should also address the state of the field. How does one currently handle graph computations in MATLAB? Are there any comparable packages in MATLAB?  If yes, what is the advantage of igraph over those packages?

Documentation/Code:
- [ ] In Getting Started -> Random Number Generators, I think "actual" and "expected" should generally not be equal if the RNG has not been reset in between calls. (The fist code block output.)
- [ ] On my (M1) mac, using the igraph RNG, I consistently get 0.19, 0.03 as the output of the following code, which is different than the output in the documentation. I worry if that affects reproducibility across OS systems:
```MATLAB
rng(1234);
x = rand(1);
igraph.cluster(g, "labelPropagation");
fprintf("%0.2f, %0.2f\n\n", x, rand(1));
```
- [ ] In Basic Usage -> Primary Functions: "... obtaining a graph, modifying graphs. characterizing graphs, plotting graphs, and saving graphs to file." a comma is needed instead of fullstop after "modifying graphs".
- [ ] In Installation -> Getting the source code, there should be an additional instruction:
git clone "https://github.com/davidrconnell/matlab-igraph"
cd matlab-igraph
git submodule update --init --recursive
```

Best,
George

related: https://github.com/openjournals/joss-reviews/issues/8622
