### 2025-08-24 ac96950
Fix RNG example to ensure clustering gives different solution
