### 2025-08-25 6da09e7
Test stochastic methods produce same results across environments
Related to #11.
