### 2025-08-28 b7a58bc
Fix typo in docs
See #11.
