### 2025-08-28 c9937d1
Add missing instruction for setting up git repo locally
See #11.
