### 2025-08-28 d28ee54
Apply suggestions from @gartavanis
See #11.
