+++ paper/paper.md
@@ -22,7 +22,7 @@ By representing graphs using MATLAB's builtin types, I maintain a simple syntax 
 
 MATLAB is a powerful tool for numerical analysis that is widely used in academia alongside Python and R, but unlike Python and R, it lacked support for igraph, a gap filled by `matlab-igraph`.
 Without `matlab-igraph`, researchers would need to rewrite algorithms specifically for MATLAB, which not only takes time but also risks errors.
-Development of `matlab-igraph` begin with work on the SpeakEasy2 community detection algorithm [@gaiteri23_robus_scalab_infor_clust_diver_biolog_networ], which needed to score partitions, using methods like modularity and normalized mutual information, and compare results to current community detection algorithms such as Leiden and InfoMAP.
+Development of `matlab-igraph` begun with work on the SpeakEasy2 community detection algorithm [@gaiteri23_robus_scalab_infor_clust_diver_biolog_networ], which needed to score partitions, using methods like modularity and normalized mutual information, and compare results to current community detection algorithms such as Leiden and InfoMAP.
 By making graph algorithms available to MATLAB, the `matlab-igraph` toolbox allows easy comparisons against state-of-the-art methods.
 This package has two goals, bring igraph functions into MATLAB for end users and ease development of new graph methods in C using the `mxIgraph` interface library and the `igutils` namespace.
 
@@ -40,9 +40,10 @@ To aid with this, the toolbox includes the bridge C library `mxIgraph` for commu
 The `mxIgraph` library exposes functions for converting between MATLAB and igraph data types, predicates for working with graphs, and a set of methods for parsing argument structures created by MATLAB's argument blocks in C.
 When large graphs are involved, C can reduce the memory demand over writing in MATLAB.
 Specifically, memory usage can be an issue in MATLAB when using "outer--reduce" patterns or parallel computation.
-For computation using matrix algebra, a common pattern is performing an outer-product like function on two vectors and then reducing over an axis.
+For computation using matrix algebra, a common pattern is computing an outer-product like function on two vectors and then reducing over an axis.
 For an $m \times{} 1$ and an $n \times{} 1$ vector, this leads to an $m \times n$ intermediary matrix that is then reduced back down to a vector with a sum along the rows or similar operation.
-By using for-loops in C instead of high-level linear algebra packages, the final vector can be created directly, without needing to store a temporary matrix, for large values of $n$ and $m$ this can be the difference between running out of memory or not.
+By using for-loops in C instead of high-level linear algebra packages, the final vector can be created directly.
+This removes the need to store an intermediary matrix and for large values of $n$ and $m$ this may be the difference between running out of memory or not.
 If the code can be written to run in parallel, C supports multithreading, in contrast to MATLAB's parallel toolbox which is based on multiprocessing, a technique that requires cloning data across each worker, potentially exhausting memory when used with large graphs, while high overhead can cause performance to be worse than serial processing for smaller graphs.
 Multithreading allows memory to be shared across threads, enabling efficient parallel processing on graphs of all sizes and does not require the MATLAB parallel toolbox.
 
