### 2025-08-28 da033f7
Apply suggestions from @gartavanis
See #11.
