### 2025-08-29 9c9712e
Provide comparison to other graph methods in matlab
