### 2025-08-29 DavidRConnell/matlab-igraph#11 @DavidRConnell
I have made the minor changes to the paper and documentation. I also added a section comparing to the builtin MATLAB graph algorithms. That should be updated on the paper branch. Let me know if this is not sufficient for addressing the state of the field.

Additionally, I added replication tests. The tests compare new randomly generated graphs to the graphs I generated on my computer (found in the [RNG tests](https://github.com/DavidRConnell/matlab-igraph/blob/master/tests/toolbox/TestRNG.m)). This tests the scenario in the livescript where the MATLAB RNG should produce the same values after being used to run the label propagation algorithm. Additionally, there's a test that ensure that each RNG produces consistent random graphs in any environment. These tests are passing for each OS used in the GitHub actions.

So I am not sure why you are seeing different values. My one guess would be that due to the limitations of running MATLAB in the actions, I am not able to run the livescripts in the actions so I have to run locally and commit the changes. It's possible changes to the toolbox impacted the values in the livescript but it was never rerun. But I was still getting the old value so I am not confident that is what happened.
