### 2025-09-02 DavidRConnell/matlab-igraph#11 @gartavanis
Thanks for making the changes and adding a state of the field section. These all look good.

Unfortunately, I'm still getting different results for that same code block: With the new changes, instead of (0.19, 0.16), I'm getting (0.19, 0.76). I've tried this with both the in-browser and the Desktop versions of MATLAB in my macbook.
