### 2025-09-03 2b654c9
Rerun livescripts when releasing versions
To prevent the scripts from becoming up of date with the code.
