+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Set RNG before first stochastic method in random number generator livescript (#11).
+
+### Changed
+
+- Update GCC from 10 -> 13 for building Linux toolbox.
+
 ## [0.2.2] 2025-08-28
 
 ### Add
