### 2025-09-04 #8622 @danielskatz
@DavidRConnell - please see https://joss.readthedocs.io/en/latest/example_paper.html and follow that example.
