### 2025-09-04 #8622 @DavidRConnell
@danielskatz the error is due to not including an affiliation.

``` text
home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.3.0/gems/theoj-1.11.1/lib/theoj/paper.rb:142:in `parse_affiliations': undefined method `each' for nil (NoMethodError)

        affiliations_yaml.each do |affiliation|
                         ^^^^^
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.3.0/gems/theoj-1.11.1/lib/theoj/paper.rb:121:in `parse_authors'
```

I removed the affiliation related to [this issue](https://github.com/DavidRConnell/matlab-igraph/issues/6), should I add a dummy affiliation or bring back the Rush affiliation?
