### 2025-09-04 ff4984a
Remove remnant affiliation from author
