+++ paper/paper.md
@@ -6,7 +6,6 @@ tags:
 authors:
   - name: David R. Connell
     orcid: 0000-0002-4841-6756
-    affiliation: 1
 date: 18 June 2025
 bibliography: paper.bib
 ---
