### 2025-09-05 edb22d5
Add affiliation back in
