### 2025-09-05 #8622 @DavidRConnell
I added the affiliation back in. If anyone would like me to handle this differently please let me know and I will change it.
