### 2025-09-05 #8622 @danielskatz
Again, see the comment I made above. In the example paper, affiliation 3 is what you should be using.
