+++ paper/paper.md
@@ -8,9 +8,8 @@ authors:
     orcid: 0000-0002-4841-6756
     affiliation: 1
 affiliations:
-  - name: Rush University Medical College, United States
+  - name: Independent Researcher, United States
     index: 1
-    ror: 01j7c0b24
 date: 18 June 2025
 bibliography: paper.bib
 ---
