### 2025-09-05 #8622 @DavidRConnell
Sorry, I was looking for an example with an empty affiliation and didn't see that.
