### 2025-09-05 d1c38c2
Update paper.md
