### 2025-09-05 dcef1d5
Update paper.md
