+++ paper/paper.md
@@ -50,7 +50,7 @@ To aid with this, the toolbox includes the bridge C library `mxIgraph` for commu
 The `mxIgraph` library exposes functions for converting between MATLAB and igraph data types, predicates for working with graphs, and a set of methods for parsing argument structures created by MATLAB's argument blocks in C.
 When large graphs are involved, C can reduce the memory demand over writing in MATLAB.
 Specifically, memory usage can be an issue in MATLAB when using "outer--reduce" patterns or parallel computation.
-For computation using matrix algebra, a common pattern is computing an outer-product like a function on two vectors and then reducing over an axis.
+For computation using matrix algebra, a common pattern is computing an outer-product like function on two vectors and then reducing over an axis.
 For an $m \times{} 1$ and an $n \times{} 1$ vector, this leads to an $m \times n$ intermediary matrix that is then reduced back down to a vector with a sum along the rows or similar operation.
 By using for-loops in C instead of high-level linear algebra packages, the final vector can be created directly.
 This removes the need to store an intermediary matrix and for large values of $n$ and $m$, this may be the difference between running out of memory or not.
