### 2025-09-05 0521052
Use more general "pairwise" term instead of "outer-product like"
