+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] 2025-09-05
+
+### Added
+
+- Dummy release to get Zenodo DOI.
+
 ## [0.2.3] 2025-09-04
 
 ### Fixed
