### 2025-09-05 #8622 @DavidRConnell
Archived version is v0.2.4 with Zenodo DOI https://doi.org/10.5281/zenodo.17064969.
