### 2025-09-05 #8622 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
