### 2025-09-05 c3e6be3
Add citation
