+++ CITATION.cff
@@ -1,0 +1,26 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Connell
+  given-names: David R.
+  orcid: "https://orcid.org/0000-0002-4841-6756"
+doi: 10.5281/zenodo.17064969
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Connell
+    given-names: David R.
+    orcid: "https://orcid.org/0000-0002-4841-6756"
+  date-published: 2025-09-05
+  doi: 10.21105/joss.08622
+  issn: 2475-9066
+  issue: 113
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 8622
+  title: "matlab-igraph: bringing igraph to MATLAB"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.08622"
+  volume: 10
+title: "matlab-igraph: bringing igraph to MATLAB"
