### 2025-09-05 9856ae7
Add JOSS badge
