### 2025-09-05 #8622 @danielskatz
Congratulations to @DavidRConnell (David R. Connell) on your JOSS publication!!

And thanks to @gartavanis and @platipodium for reviewing!
JOSS fully depends on volunteers and we couldn't do this without you
